Implement feedback on PR #2821

+++ PythonClient/airsim/types.py
@@ -415,9 +415,9 @@ class GpsData(MsgpackMixin):
 
 class DistanceSensorData(MsgpackMixin):
     time_stamp = np.uint64(0)
-    distance = np.float32(0)
-    min_distance = np.float32(0)
-    max_distance = np.float32(0)
+    distance = 0.0
+    min_distance = 0.0
+    max_distance = 0.0
     relative_pose = Pose()
 
 class PIDGains():
